Introduced sbp_signal_t into use.

+++ src/acq.h
@@ -23,6 +23,6 @@ void acq_service_load_done(void);
 void acq_search(float cf_min, float cf_max, float cf_bin_width);
 void acq_service_irq(void);
 void acq_get_results(float* cp, float* cf, float* cn0);
-void acq_send_result(u8 prn, float snr, float cp, float cf);
+void acq_send_result(u16 prn, float snr, float cp, float cf);
 
 #endif
